added changes to .env file and moved fetchLocation function to seperate folder
