+++ package.json
@@ -14,13 +14,12 @@
 		"leaflet": "^1.9.3",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"react-dotenv": "^0.1.3",
 		"react-leaflet": "^4.2.1",
 		"react-scripts": "5.0.1",
 		"web-vitals": "^2.1.4"
 	},
 	"scripts": {
-		"start": "react-dotenv && react-scripts start",
+		"start": "react-scripts start",
 		"build": "react-scripts build",
 		"test": "react-scripts test",
 		"eject": "react-scripts eject",
@@ -53,10 +52,5 @@
 		"prettier": "2.8.4",
 		"tailwindcss": "^3.2.7",
 		"typescript": "^4.9.5"
-	},
-	"react-dotenv": {
-		"whitelist": [
-			"REACT_APP_KEY"
-		]
 	}
 }
