--- src/.env
@@ -1,1 +1,0 @@
-REACT_APP_KEY=at_5I9sUYamw5MsQycwQTbrB6gMIRezB
