+++ src/api/apiFetchLocation.tsx
@@ -1,0 +1,18 @@
+import api_baseUrl from "./api_baseUrl";
+
+export const useFetchLocation = (ipAddress: string) => {
+	const fetchLocation = async () => {
+		try {
+			const response = await api_baseUrl.get(
+				`/country?apiKey=${process.env.REACT_APP_KEY}&ipAddress=${ipAddress}`
+			);
+			console.log(response.data);
+		} catch (err: any) {
+			console.log(err.response.data);
+			console.log(err.response.status);
+		}
+	};
+	fetchLocation();
+
+	return { fetchLocation };
+};
